Move consensus logic into system & aura.
