+++ Cargo.toml
@@ -62,7 +62,6 @@ members = [
 	"srml/assets",
 	"srml/aura",
 	"srml/balances",
-	"srml/consensus",
 	"srml/contract",
 	"srml/council",
 	"srml/democracy",
