--- srml/consensus/Cargo.toml
@@ -1,31 +1,0 @@
-[package]
-name = "srml-consensus"
-version = "2.0.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-edition = "2018"
-
-[dependencies]
-serde = { version = "1.0", optional = true, features = ["derive"] }
-parity-codec = { version = "3.3", default-features = false, features = ["derive"] }
-substrate-primitives = { path = "../../core/primitives", default-features = false }
-inherents = { package = "substrate-inherents", path = "../../core/inherents", default-features = false }
-rstd = { package = "sr-std", path = "../../core/sr-std", default-features = false }
-primitives = { package = "sr-primitives", path = "../../core/sr-primitives", default-features = false }
-srml-support = { path = "../support", default-features = false }
-system = { package = "srml-system", path = "../system", default-features = false }
-
-[dev-dependencies]
-runtime_io = { package = "sr-io", path = "../../core/sr-io" }
-
-[features]
-default = ["std"]
-std = [
-	"serde",
-	"parity-codec/std",
-	"substrate-primitives/std",
-	"rstd/std",
-	"srml-support/std",
-	"primitives/std",
-	"system/std",
-	"inherents/std",
-]
